--- src/main/java/chess/controller/DoubleDTO.java
@@ -1,13 +1,0 @@
-package chess.controller;
-
-public class DoubleDTO {
-    private final double value;
-
-    public DoubleDTO(double value) {
-        this.value = value;
-    }
-
-    public double getValue() {
-        return value;
-    }
-}
