+++ src/main/java/chess/dto/BoardDTO.java
@@ -1,7 +1,9 @@
-package chess.controller;
+package chess.dto;
 
 import chess.domain.location.Position;
 import chess.domain.piece.Piece;
+import chess.dto.PieceDTO;
+import chess.dto.PositionDTO;
 
 import java.util.HashMap;
 import java.util.Map;
