+++ src/main/java/chess/dto/ErrorDTO.java
@@ -1,4 +1,4 @@
-package chess.controller;
+package chess.dto;
 
 public class ErrorDTO {
     private final String message;
